+++ api/package.json
@@ -15,7 +15,7 @@
     "dev": "ERXES_TELEMETRY_DISABLED=1 DEBUG=erxes* NODE_ENV=development node_modules/.bin/ts-node-dev --respawn --ignore-watch node_modules -- src",
     "dev-crons": "PROCESS_NAME=crons DEBUG=erxes-crons:* NODE_ENV=development node_modules/.bin/ts-node-dev --respawn src/cronJobs",
     "dev-workers": "PROCESS_NAME=workers DEBUG=erxes-workers:* NODE_ENV=development NODE_OPTIONS='--experimental-worker' node_modules/.bin/ts-node src/workers/index.ts",
-    "build": "tsc -p tsconfig.prod.json && cp -rf src/private dist/ && cp -rf src/initialData dist/",
+    "build": "tsc -p tsconfig.prod.json && cp -rf src/private dist/ && cp -rf src/initialData dist/ && mv ./dist/api/src/* ./dist",
     "test": "node --expose-gc --max_old_space_size=4000 ./node_modules/.bin/jest --forceExit --silent --logHeapUsage --w=4"
   },
   "dependencies": {
