+++ api/tsconfig.prod.json
@@ -1,10 +1,17 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
+    "outDir": "./dist/api/src",
     "sourceMap": false,
     "inlineSourceMap": false,
-    "inlineSources": false
+    "inlineSources": false,
+    "strictNullChecks": false
   },
-  "include": ["./src/**/*.ts", "./src/**/*.js"],
+  "include": [
+    "./src/**/*.ts",
+    "./src/**/*.js",
+    "typings.d.ts",
+    "../plugins/**/api"
+  ],
   "exclude": ["node_modules", "dist", "src/__tests__"]
 }
